+++ src/joseph/peterson/myNewHibernate/src/PhoneNumber.java
@@ -1,0 +1,21 @@
+package joseph.peterson.myNewHibernate.src;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Table;
+@Entity
+@Table(name = "phone_number")
+public class PhoneNumber {
+
+    @Id
+    @GeneratedValue
+    private Integer id;
+    private String phone;
+    public String getPhone() {
+        return phone;
+    }
+    public void setPhone(String phone) {
+        this.phone = phone;
+    }
+}
